Implement filter buttons
- Add filter buttons (none, all, user). These select the set of playlists as described, and update playlist selection. Then view is refreshed to reflect the selection changes.
- Add overwrite export mode, and allow user to switch been overwrite and append export mode

+++ mpfree/UI/WindowController.swift
@@ -8,28 +8,41 @@
 
 import Cocoa
 
+enum ExportMode: Int {
+    case append = 0
+    case overwrite = 1
+}
+
 class WindowController: NSWindowController {
     
     var appDelegate: AppDelegate!
     var parser: ITParse!
+    var exportOverwrite: ExportMode! = .append
     
-    @IBAction func doExportAdd(_ sender: NSToolbarItem) {
-        let panel = NSOpenPanel()
-        panel.canChooseFiles = false
-        panel.canChooseDirectories = true
-        panel.allowsMultipleSelection = false
-        
-        panel.message = "Select a location to export the playlists to."
-        
-        let response = panel.runModal()
-        
-        if response == NSApplication.ModalResponse.OK {
-            guard let selectedURL = panel.url else { return }
-            parser.exportPlaylists(exportURL: selectedURL)
-        }
+    @IBAction func selectExportMode(_ sender: Any) {
+        // Set a new export mode when the user clicks on the segmented control.
+        guard let segmentedControl = sender as? NSSegmentedControl else { return }
+        exportOverwrite = ExportMode.init(rawValue: segmentedControl.selectedSegment)
+        print("WindowController.selectExportMode: \(exportOverwrite!) export mode selected")
     }
     
-    @IBAction func doExportOverwrite(_ sender: NSToolbarItem) {
+    @IBAction func selectFilterCriteria(_ sender: Any) {
+        // Apply a set of filters to the playlists depending on which button was pressed.
+        guard let segmentedControl = sender as? NSSegmentedControl else { return }
+        let selectionState: SelectionState! = SelectionState.init(rawValue: segmentedControl.selectedSegment)
+        print("WindowController.selectFilterCriteria: \(String(describing: selectionState)) filter criteria selected")
+        
+        // Update the selection status of the playlist model.
+        parser.targetTree.setSelection(selectionState)
+        
+        // Redraw the view.
+        parser.redrawTree()
+    }
+    
+    @IBAction func selectExport(_ sender: NSToolbarItem) {
+        // Perform the playlist export.
+        
+        print("WindowController.selectExport: export selected")
         
         let panel = NSOpenPanel()
         panel.canChooseFiles = false
@@ -43,35 +56,44 @@ class WindowController: NSWindowController {
         if response == NSApplication.ModalResponse.OK {
             guard let selectedURL = panel.url else { return }
             
+            print("WindowController.selectExport: user has selected following path:")
             print(selectedURL.path)
             
-            if FileManager.default.fileExists(atPath: selectedURL.path) {
+            if exportOverwrite == .overwrite {
+                // If overwrite mode, remove the folder first.
                 
-                print("file exists")
-                
-                
-                // Remove the folder first.
+                print("WindowController.doExport: Clearing folder")
+                if FileManager.default.fileExists(atPath: selectedURL.path) {
+                    
+                    do {
+                        try FileManager.default.removeItem(atPath: selectedURL.path)
+                    } catch {
+                        print(error.localizedDescription);
+                    }
+                    
+                    if FileManager.default.fileExists(atPath: selectedURL.path) {
+                        print("WindowController.doExport: Failed to remove folder")
+                    }
+                }
+            }
+            
+
+            if !FileManager.default.fileExists(atPath: selectedURL.path) {
+                // If folder does not exist, make folder.
+
+                print("WindowController.doExport: Creating new folder.")
                 do {
-                    try FileManager.default.removeItem(atPath: selectedURL.path)
+                    try FileManager.default.createDirectory(atPath: selectedURL.path, withIntermediateDirectories: true, attributes: nil)
                 } catch {
                     print(error.localizedDescription);
                 }
                 
-                if FileManager.default.fileExists(atPath: selectedURL.path) {
-                    print("Failed to remove folder.")
-                }
-                
             }
             
-            
-            // Make a new folder in the same place.
-            do {
-                try FileManager.default.createDirectory(atPath: selectedURL.path, withIntermediateDirectories: true, attributes: nil)
-            } catch {
-                print(error.localizedDescription);
-            }
-            
+            // Call export function.
             parser.exportPlaylists(exportURL: selectedURL)
+        } else {
+            print("WindowController.selectExport: user cancelled")
         }
     }
     
@@ -80,8 +102,6 @@ class WindowController: NSWindowController {
         super.windowDidLoad()
         
         // Implement this method to handle any initialization after your window controller's window has been loaded from its nib file.
-        
-        
         appDelegate = (NSApplication.shared.delegate as! AppDelegate)
         parser = appDelegate.parser
